+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.3
+
+Added persistant storage. When a logged in user closes the app, then reopens, the data
+will now be persistant. Best practice would be to Parse.init, then Parse.currentUser. This
+will return the current user session and allow autologin.
+
 ## 1.0.2
 
 Fixed login